REWORKED PROJECT LAYOUT AND ADDED BASIC AUDIO PLAYBACK
The project is now split into several logical modules.

Audio playback is added using the command line tool mpg321 running in an asyncio subprocess.

Also found something strange with the autonomous driving code. For some reason the comparison of the motor task duration and the current time is not evaluating as I expect. Check the TODO code in robot.py